requirement.txt
absl-py==2.0.0
alembic==1.13.1
aniso8601==9.0.1
apispec==5.2.2
apispec-webframeworks==0.5.2
arrow==1.3.0
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
bidict==0.22.1
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
constantly==23.10.4
cookiecutter==2.1.1
dask==2023.12.1
deepdiff==6.2.3
distributed==2023.12.1
dnspython==2.4.2
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fsspec==2023.12.2
gast==0.5.4
gevent==23.9.1
gevent-websocket==0.10.1
gitignore_parser==0.1.9
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
keras==2.15.0
kthread==0.2.3
libclang==16.0.6
locket==1.0.0
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.28.1
ml-dtypes==0.2.0
modin==0.26.0
msgpack==1.0.7
networkx==2.8.8
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
partd==1.4.1
passlib==1.7.4
Pillow==10.1.0
protobuf==4.23.4
psutil==5.9.7
pyarrow==10.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-slugify==8.0.1
python-socketio==5.10.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simple-websocket==0.10.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
taipy==3.0.0
taipy-config==3.0.0
taipy-core==3.0.0
taipy-gui==3.0.0
taipy-rest==3.0.0
taipy-templates==3.0.0
tblib==3.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
Twisted==23.10.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
zict==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
